(demo) Add a manual run github action to create RPMs for a PR  (#8)
* (demo) Update spec to 2.2.0-4

  - 2.2.0-4: Actions Changes

  - 2.2.0-3: Update spec with necessary BuildRequires

    Add packages required by the automation/*.sh scripts as `BuildRequires`
    entries in the spec file.  They will be picked up by ovirt's build
    container from there and the scripts and will be able to run without
    additional `dnf install` commands.

* (demo) Add a manual run github action to create RPMs for a PR

To support the use case of testing a rpm for changes in a specific PR,
add a github action that runs when a comment "/build rpm" is created
on a PR.

When the build is completed successfully, a comment with a link to the
build action's summary is added for easy access.  The rpm repo can be
downloaded from the link.

This action will run the build in the same way as copr runs a build:
  - setup the container (packages, sources)
  - create the source rpm
  - install any build requirements for the source rpm
  - build the rpm from the source rpm
  - publish a rpm repo of the results

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>